+++ examples/src/json_convert_examples.cpp
@@ -111,12 +111,12 @@ void streaming_example2()
 
     json_serializer serializer(std::cout, jsoncons::indenting::indent); 
 
-    serializer.begin_json();       
+    serializer.begin_document();       
     serializer.begin_object();       
     serializer.name("Employees");       
     encode_fragment(employees, serializer);
     serializer.end_object();       
-    serializer.end_json();       
+    serializer.end_document();       
 
     std::cout << "\n\n";
 }