+++ examples/src/serialization_examples.cpp
@@ -297,7 +297,7 @@ void dump_json_fragments()
     )");
 
     json_serializer serializer(std::cout, jsoncons::indenting::indent); // pretty print
-    serializer.begin_json();
+    serializer.begin_document();
     serializer.begin_array();
     for (const auto& book : some_books.array_range())
     {
@@ -308,7 +308,7 @@ void dump_json_fragments()
         book.dump_fragment(serializer);
     }
     serializer.end_array();
-    serializer.end_json();
+    serializer.end_document();
 }
 
 void nan_inf_replacement()