+++ include/jsoncons/json_convert_traits.hpp
@@ -69,14 +69,14 @@ struct json_convert_traits<T,
     template <class CharT>
     static void encode(const T& val, basic_json_content_handler<CharT>& serializer)
     {
-        serializer.begin_json();
+        serializer.begin_document();
         serializer.begin_array();
         for (auto it = std::begin(val); it != std::end(val); ++it)
         {
             json_convert_traits<value_type>::encode(*it,serializer);
         }
         serializer.end_array();
-        serializer.end_json();
+        serializer.end_document();
     }
 };
 
@@ -105,14 +105,14 @@ struct json_convert_traits<std::array<T,N>>
     template <class CharT>
     static void encode(const std::array<T, N>& val, basic_json_content_handler<CharT>& serializer)
     {
-        serializer.begin_json();
+        serializer.begin_document();
         serializer.begin_array();
         for (auto it = std::begin(val); it != std::end(val); ++it)
         {
             json_convert_traits<value_type>::encode(*it,serializer);
         }
         serializer.end_array();
-        serializer.end_json();
+        serializer.end_document();
     }
 };
 
@@ -137,7 +137,7 @@ struct json_convert_traits<T,
     template <class CharT>
     static void encode(const T& val, basic_json_content_handler<CharT>& serializer)
     {
-        serializer.begin_json();
+        serializer.begin_document();
         serializer.begin_object();
         for (auto it = std::begin(val); it != std::end(val); ++it)
         {
@@ -145,7 +145,7 @@ struct json_convert_traits<T,
             json_convert_traits<mapped_type>::encode(it->second,serializer);
         }
         serializer.end_object();
-        serializer.end_json();
+        serializer.end_document();
     }
 };
 
@@ -237,9 +237,9 @@ T decode_json(std::basic_istringstream<CharT>& is,
 template <class T, class CharT>
 void encode_json(const T& val, basic_json_content_handler<CharT>& handler)
 {
-    handler.begin_json();
+    handler.begin_document();
     json_convert_traits<T>::encode(val,handler);
-    handler.end_json();
+    handler.end_document();
 }
 
 template <class T, class CharT>